feat: define shared AnalyticsEvent
